+++ mc2/controllers/docker/tests/test_docker_controller.py
@@ -2,8 +2,12 @@ import pytest
 import responses
 from django.conf import settings
 from django.contrib.auth.models import User
+from django.core.urlresolvers import reverse
+
 from mc2.controllers.base.tests.base import ControllerBaseTestCase
-from mc2.controllers.docker.models import DockerController, traefik_domains
+from mc2.controllers.docker.models import (
+    DockerController, marathon_lb_domains, traefik_domains)
+from mc2.organizations.models import Organization, OrganizationUserRelation
 
 
 @pytest.mark.django_db
@@ -21,6 +25,11 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             marathon_cmd='ping',
             docker_image='docker/image',
         )
+        org = Organization.objects.create(name="Test Org", slug="test-org")
+        OrganizationUserRelation.objects.create(
+            user=self.user, organization=org)
+        controller.organization = org
+        controller.save()
 
         custom_urls = "testing.com url.com"
         controller.domain_urls += custom_urls
@@ -37,9 +46,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "test-org",
             },
             "container": {
                 "type": "DOCKER",
@@ -47,8 +57,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "image": "docker/image",
                     "forcePullImage": True,
                     "network": "BRIDGE",
-                }
-            }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
+            },
         })
 
         controller.port = 1234
@@ -67,9 +78,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
-                "name": "Test App"
+                "name": "Test App",
+                "org": "test-org",
             },
             "container": {
                 "type": "DOCKER",
@@ -78,7 +90,8 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "forcePullImage": True,
                     "network": "BRIDGE",
                     "portMappings": [{"containerPort": 1234, "hostPort": 0}],
-                }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
             },
         })
 
@@ -98,9 +111,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "test-org",
             },
             "container": {
                 "type": "DOCKER",
@@ -109,6 +123,7 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "forcePullImage": True,
                     "network": "BRIDGE",
                     "portMappings": [{"containerPort": 1234, "hostPort": 0}],
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
                 }
             },
             "ports": [0],
@@ -140,9 +155,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "test-org",
             },
             "container": {
                 "type": "DOCKER",
@@ -152,6 +168,7 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "network": "BRIDGE",
                     "portMappings": [{"containerPort": 1234, "hostPort": 0}],
                     "parameters": [
+                        {"key": "memory-swappiness", "value": "0"},
                         {"key": "volume-driver", "value": "xylem"},
                         {
                             "key": "volume",
@@ -188,9 +205,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "test-org",
             },
             "container": {
                 "type": "DOCKER",
@@ -200,6 +218,7 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "network": "BRIDGE",
                     "portMappings": [{"containerPort": 1234, "hostPort": 0}],
                     "parameters": [
+                        {"key": "memory-swappiness", "value": "0"},
                         {"key": "volume-driver", "value": "xylem"},
                         {
                             "key": "volume",
@@ -244,9 +263,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "",
             },
             "container": {
                 "type": "DOCKER",
@@ -254,8 +274,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "image": "docker/image",
                     "forcePullImage": True,
                     "network": "BRIDGE",
-                }
-            }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
+            },
         })
 
     def test_get_marathon_app_data_with_app_labels(self):
@@ -281,10 +302,11 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
-                "TEST_LABELS_NAME": 'a test label value'
+                "TEST_LABELS_NAME": 'a test label value',
+                "org": "",
             },
             "container": {
                 "type": "DOCKER",
@@ -292,8 +314,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "image": "docker/image",
                     "forcePullImage": True,
                     "network": "BRIDGE",
-                }
-            }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
+            },
         })
 
     @responses.activate
@@ -327,9 +350,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
-                "name": "Test App"
+                "name": "Test App",
+                "org": "",
             },
             "container": {
                 "type": "DOCKER",
@@ -337,8 +361,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "image": "docker/image",
                     "forcePullImage": True,
                     "network": "BRIDGE",
-                }
-            }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
+            },
         })
 
     @responses.activate
@@ -381,9 +406,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
             "labels": {
                 "domain": domain_label,
                 "HAPROXY_GROUP": "external",
-                "HAPROXY_0_VHOST": domain_label,
+                "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                 "traefik.frontend.rule": traefik_domains(domain_label),
                 "name": "Test App",
+                "org": "",
             },
             "container": {
                 "type": "DOCKER",
@@ -391,8 +417,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "image": "docker/image",
                     "forcePullImage": True,
                     "network": "BRIDGE",
-                }
-            }
+                    "parameters": [{"key": "memory-swappiness", "value": "0"}],
+                },
+            },
         })
 
     @responses.activate
@@ -425,9 +452,10 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                 "labels": {
                     "domain": domain_label,
                     "HAPROXY_GROUP": "external",
-                    "HAPROXY_0_VHOST": domain_label,
+                    "HAPROXY_0_VHOST": marathon_lb_domains(domain_label),
                     "traefik.frontend.rule": traefik_domains(domain_label),
                     "name": "Test App",
+                    "org": "",
                 },
                 "container": {
                     "type": "DOCKER",
@@ -437,7 +465,9 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                         "network": "BRIDGE",
                         "portMappings": [
                             {"containerPort": 1234, "hostPort": 0}],
-                    }
+                        "parameters": [
+                            {"key": "memory-swappiness", "value": "0"}],
+                    },
                 },
                 "ports": [0],
                 "healthChecks": [{
@@ -450,3 +480,50 @@ class DockerControllerTestCase(ControllerBaseTestCase):
                     "timeoutSeconds": 200
                 }]
             })
+
+    @responses.activate
+    def test_create_new_controller_with_no_port(self):
+        org = Organization.objects.create(name="Foo Org", slug="foo-org")
+        OrganizationUserRelation.objects.create(
+            user=self.user, organization=org)
+
+        self.client.login(username='testuser2', password='test')
+        self.client.get(
+            reverse('organizations:select-active', args=('foo-org',)))
+
+        self.mock_create_marathon_app()
+        self.mock_create_postgres_db(200, {
+            'result': {
+                'name': 'joes_db',
+                'user': 'joe',
+                'password': '1234',
+                'host': 'localhost'}})
+
+        data = {
+            'name': 'Another test app',
+            'docker_image': 'test/image',
+            'postgres_db_needed': True,
+            'env-TOTAL_FORMS': 0,
+            'env-INITIAL_FORMS': 0,
+            'env-MIN_NUM_FORMS': 0,
+            'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
+            'link-TOTAL_FORMS': 0,
+            'link-INITIAL_FORMS': 0,
+            'link-MIN_NUM_FORMS': 0,
+            'link-MAX_NUM_FORMS': 100,
+
+
+        }
+
+        response = self.client.post(reverse('controllers.docker:add'), data)
+        self.assertEqual(response.status_code, 302)
+
+        controller = DockerController.objects.all().last()
+        self.assertEqual(controller.state, 'done')
+        self.assertEqual(controller.name, 'Another test app')
+        self.assertEqual(controller.organization.slug, 'foo-org')
+        self.assertIsNone(controller.port)