+++ mc2/controllers/docker/tests/test_docker_controller_hypothesis.py
@@ -1,8 +1,9 @@
 import json
-import string
-
+import mock
 import pytest
 import responses
+import string
+
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
@@ -14,7 +15,9 @@ from hypothesis.strategies import text, random_module, lists, just
 
 from mc2.controllers.base.models import EnvVariable, MarathonLabel
 from mc2.controllers.base.tests.base import ControllerBaseTestCase
-from mc2.controllers.docker.models import DockerController, traefik_domains
+from mc2.controllers.docker.models import (
+    DockerController, marathon_lb_domains, traefik_domains)
+from mc2.controllers.base.managers.rabbitmq import ControllerRabbitMQManager
 from mc2.organizations.models import Organization
 
 
@@ -57,6 +60,14 @@ def docker_controller(with_envvars=True, with_labels=True, **kw):
     domain_urls = text(string.ascii_letters + string.digits + '-.')
     kw.setdefault("domain_urls", domain_urls)
 
+    # Prevent Hypothesis from generating vhosts with invalid characters
+    rabbitmq_vhost_name = text(string.ascii_letters + string.digits + '-.')
+    kw.setdefault("rabbitmq_vhost_name", rabbitmq_vhost_name)
+
+    # Prevent Hypothesis from generating usernames with invalid characters
+    rabbitmq_vhost_username = text(string.ascii_letters + string.digits + '-.')
+    kw.setdefault("rabbitmq_vhost_username", rabbitmq_vhost_username)
+
     # The model generator sees `controller_ptr` (from the PolymorphicModel
     # magic) as a mandatory field and objects if we don't provide a value for
     # it.
@@ -92,6 +103,30 @@ def check_and_clear_appdata(appdata, controller):
     assert appdata == {}
 
 
+def check_and_remove_docker_param(parameters, key, value):
+    """
+    Assert that a particular docker param is correct and remove it.
+    """
+    paraml = [p for p in parameters if p["key"] == key]
+    assert paraml == [{"key": key, "value": value}]
+    parameters.remove(paraml[0])
+
+
+def check_and_remove_docker_params(docker, controller):
+    """
+    Assert that the docker params data is correct and remove it.
+    """
+    parameters = docker.pop("parameters")
+    check_and_remove_docker_param(parameters, "memory-swappiness", "0")
+    if controller.volume_needed:
+        check_and_remove_docker_param(parameters, "volume-driver", "xylem")
+        volume = u"%s_media:%s" % (
+            controller.app_id,
+            controller.volume_path or settings.MARATHON_DEFAULT_VOLUME_PATH)
+        check_and_remove_docker_param(parameters, "volume", volume)
+    assert parameters == []
+
+
 def check_and_remove_docker(appdata, controller):
     """
     Assert that the docker container data is correct and remove it.
@@ -105,14 +140,7 @@ def check_and_remove_docker(appdata, controller):
     if controller.port:
         assert docker.pop("portMappings") == [
             {"containerPort": controller.port, "hostPort": 0}]
-    if controller.volume_needed:
-        volume = u"%s_media:%s" % (
-            controller.app_id,
-            controller.volume_path or settings.MARATHON_DEFAULT_VOLUME_PATH)
-        assert sorted(docker.pop("parameters")) == sorted([
-            {"key": "volume-driver", "value": "xylem"},
-            {"key": "volume", "value": volume},
-        ])
+    check_and_remove_docker_params(docker, controller)
     assert docker == {}
     assert container == {}
 
@@ -146,13 +174,21 @@ def check_and_remove_env(appdata, controller):
     Assert that the correct envvars are in appdata and remove them.
     """
     env_variables = controller.env_variables.all()
-    if env_variables or controller.postgres_db_needed:
+    if env_variables or controller.postgres_db_needed or \
+            (controller.rabbitmq_vhost_needed and
+             controller.rabbitmq_vhost_name):
         envs = appdata.pop("env")
 
         if controller.postgres_db_needed:
             assert envs.pop(
                 "DATABASE_URL") == 'postgres://trevor:1234@localhost/trevordb'
 
+        if controller.rabbitmq_vhost_needed and controller.rabbitmq_vhost_name:
+            assert envs.pop("BROKER_URL") == 'amqp://%s:%s@%s/%s' % (
+                controller.rabbitmq_vhost_username,
+                controller.rabbitmq_vhost_password,
+                controller.rabbitmq_vhost_host,
+                controller.rabbitmq_vhost_name)
         if env_variables:
             # We may have duplicate keys in here, but hopefully the database
             # always return the objects in the same order.
@@ -167,16 +203,23 @@ def check_and_remove_labels(appdata, controller):
     """
     labels = appdata.pop("labels")
     assert labels.pop("name") == controller.name
+    assert labels.pop("org") == controller.organization.slug \
+        if controller.organization else ''
     domains = [u".".join([controller.app_id, settings.HUB_DOMAIN])]
     domains.extend(controller.domain_urls.split())
-    assert sorted(labels.pop("domain").split()) == sorted(domains)
-    assert sorted(labels.pop("HAPROXY_0_VHOST").split()) == sorted(domains)
-    assert labels.pop("HAPROXY_GROUP") == "external"
+    assert sorted(labels.pop("HAPROXY_0_VHOST").split(",")) == sorted(domains)
 
-    traefik_domains = labels.pop("traefik.frontend.rule")
-    traefik_domains = traefik_domains.split(":", 2)[-1].split(",")
-    traefik_domains = [d.strip() for d in traefik_domains]
-    assert sorted(traefik_domains) == sorted(domains)
+    haproxy_group = labels.pop("HAPROXY_GROUP")
+
+    if controller.external_visibility:
+        assert sorted(labels.pop("domain").split()) == sorted(domains)
+        traefik_domains = labels.pop("traefik.frontend.rule")
+        traefik_domains = traefik_domains.split(":", 2)[-1].split(",")
+        traefik_domains = [d.strip() for d in traefik_domains]
+        assert sorted(traefik_domains) == sorted(domains)
+        assert haproxy_group == "external"
+    else:
+        assert haproxy_group == "internal"
 
     # We may have duplicate keys in here, but hopefully the database always
     # return the objects in the same order.
@@ -191,6 +234,24 @@ def check_and_remove_optional(appdata, field, value):
     if value:
         assert appdata.pop(field) == value
     assert field not in appdata
+
+
+def test_marathon_lb_domains_single():
+    """
+    When a domains string with a single domain is passed to
+    marathon_lb_domains, it returns the domain unchanged.
+    """
+    domains = 'abc.com'
+    assert marathon_lb_domains(domains) == 'abc.com'
+
+
+def test_marathon_lb_domains_multiple():
+    """
+    When a domains string with multiple domains is passed to
+    marathon_lb_domains, it returns a comma-separated list of domains.
+    """
+    domains = 'abc.com def.co.za   ghi.co.ng'
+    assert marathon_lb_domains(domains) == 'abc.com,def.co.za,ghi.co.ng'
 
 
 def test_traefik_domains_single():
@@ -239,32 +300,47 @@ class DockerControllerHypothesisTestCase(TestCase, ControllerBaseTestCase):
                 'host': 'localhost'}})
 
     @responses.activate
+    @mock.patch.object(ControllerRabbitMQManager, '_create_username')
     @hsettings(perform_health_check=False)
     @given(_r=random_module(), controller=docker_controller())
-    def test_get_marathon_app_data(self, _r, controller):
+    def test_get_marathon_app_data(self, _r, controller, mock_create_u):
         """
         Suitable app_data is built for any combination of model parameters.
         """
+        if controller.rabbitmq_vhost_needed and controller.rabbitmq_vhost_name:
+            self.mock_successful_new_vhost(
+                controller.rabbitmq_vhost_name,
+                controller.rabbitmq_vhost_username)
+            mock_create_u.return_value = controller.rabbitmq_vhost_username
+
         app_data = controller.get_marathon_app_data()
         check_and_clear_appdata(app_data, controller)
 
     @responses.activate
+    @mock.patch.object(ControllerRabbitMQManager, '_create_username')
     @hsettings(perform_health_check=False)
     @given(_r=random_module(), controller=docker_controller())
-    def test_from_marathon_app_data(self, _r, controller):
+    def test_from_marathon_app_data(self, _r, controller, u):
         """
         A model imported from app_data generates the same app_data as the model
         it was imported from.
         """
+        if controller.rabbitmq_vhost_needed and controller.rabbitmq_vhost_name:
+            self.mock_successful_new_vhost(
+                controller.rabbitmq_vhost_name,
+                controller.rabbitmq_vhost_username)
+            u.return_value = controller.rabbitmq_vhost_username
+
         app_data = controller.get_marathon_app_data()
         new_controller = DockerController.from_marathon_app_data(
             controller.owner, controller.organization, app_data)
         assert app_data == new_controller.get_marathon_app_data()
 
     @responses.activate
+    @mock.patch.object(ControllerRabbitMQManager, '_create_username')
     @hsettings(perform_health_check=False, max_examples=50)
     @given(_r=random_module(), controller=docker_controller(), name=text())
-    def test_from_marathon_app_data_with_name(self, _r, controller, name):
+    def test_from_marathon_app_data_with_name(self, _r, controller, name, u):
         """
         A model imported from app_data generates the same app_data as the model
         it was imported from, but with the name field overridden.
@@ -272,6 +348,12 @@ class DockerControllerHypothesisTestCase(TestCase, ControllerBaseTestCase):
         We limit the number of examples we generate because we don't need
         hundreds of examples to verify that this behaviour is correct.
         """
+        if controller.rabbitmq_vhost_needed and controller.rabbitmq_vhost_name:
+            self.mock_successful_new_vhost(
+                controller.rabbitmq_vhost_name,
+                controller.rabbitmq_vhost_username)
+            u.return_value = controller.rabbitmq_vhost_username
+
         app_data = controller.get_marathon_app_data()
         new_controller = DockerController.from_marathon_app_data(
             controller.owner, controller.organization, app_data, name=name)
@@ -282,9 +364,10 @@ class DockerControllerHypothesisTestCase(TestCase, ControllerBaseTestCase):
         assert app_data_with_name == new_controller.get_marathon_app_data()
 
     @responses.activate
+    @mock.patch.object(ControllerRabbitMQManager, '_create_username')
     @hsettings(perform_health_check=False, max_examples=50)
     @given(_r=random_module(), controller=docker_controller(), name=text())
-    def test_hidden_import_view(self, _r, controller, name):
+    def test_hidden_import_view(self, _r, controller, name, mock_create_u):
         """
         A model imported through the hidden view generates the same app_data
         as the model it was imported from, but with the name field overridden.
@@ -292,6 +375,12 @@ class DockerControllerHypothesisTestCase(TestCase, ControllerBaseTestCase):
         We limit the number of examples we generate because we don't need
         hundreds of examples to verify that this behaviour is correct.
         """
+        if controller.rabbitmq_vhost_needed and controller.rabbitmq_vhost_name:
+            self.mock_successful_new_vhost(
+                controller.rabbitmq_vhost_name,
+                controller.rabbitmq_vhost_username)
+            mock_create_u.return_value = controller.rabbitmq_vhost_username
+
         app_data = controller.get_marathon_app_data()
         user = controller.owner
         user.set_password("password")  # So we can log in.