+++ mc2/urls.py
@@ -13,7 +13,14 @@ urlpatterns = patterns(
         views.HomepageView.as_view(),
         name='home'
     ),
+    url(
+        r'^dashboard/$',
+        views.DashboardView.as_view(),
+        name='dashboard'
+    ),
     url(r'^login/?$', views.MC2LoginView.as_view(), name='login'),
+    url(r'^create_account/?$', views.CreateAccountView.as_view(),
+        name='create_account'),
     url(r'', include('mama_cas.urls')),
     url(
         r'^logout/$',